requirements.txt
slackclient==2.9.3
firebase_admin==4.5.3
